+++ src/main/java/org/osgl/exception/UnexpectedInterruptedException.java
@@ -1,0 +1,14 @@
+package org.osgl.exception;
+
+/**
+ * This is a RuntimeException version of {@link InterruptedException}
+ */
+public class UnexpectedInterruptedException extends UnexpectedException {
+    public UnexpectedInterruptedException(InterruptedException cause) {
+        super(cause);
+    }
+
+    public InterruptedException toInterruptedException() {
+        return (InterruptedException) getCause();
+    }
+}